requirements.txt
psycopg2-binary
numpy
sentence-transformers
